+++ src/main/java/lagatrix/exceptions/FileException.java
@@ -1,0 +1,15 @@
+package lagatrix.exceptions;
+
+/**
+ * This exception is used in file exceptions.
+ *
+ * @author javierfh03
+ * @since 0.3
+ */
+public class FileException extends LagatrixException {
+
+    public FileException(String description) {
+        super(description);
+    }
+
+}